Correct wantedby target
This causes lernstick-quirks to be executed only *once*. Else it is executed before systemd-udevd.service AND as sysinit.target dependency. This is specially needed in lernstick-exam-client in the warm reboot. Else, this causes the brcmfmac driver to be unloaded, resulting in a vanishing network card on certain MacBook Pro systems. This service needs to run *only* before systemd-udevd.service.